# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: EXPP API
  version: "1.0.0"
  description: >
    EXPP — Educational Workflow Platform v1.
    All endpoints require authentication via the `expp_session` HTTP-only cookie
    unless noted as public.
  contact:
    name: EXPP Project
servers:
  - url: /api
    description: Next.js App Router API
...
components:
  schemas:
    ApiSuccess:
      type: object
      required: [success, data, traceId]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          description: Response payload (schema varies per endpoint)
        traceId:
          type: string
    ApiError:
      type: object
      required: [success, error, code, traceId]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string
          enum: [AUTH_ERROR, FORBIDDEN, NOT_FOUND, VALIDATION_ERROR, RATE_LIMITED]
        traceId:
          type: string
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: expp_session
security:
  - sessionCookie: []

# ── Paths ──────────────────────────────────────────────────────────────────
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Authenticated. Session cookie set.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "429":
          description: Rate limited (10 req/min per IP)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /generation/{id}/regenerate:
    post:
      tags: [generation]
      summary: Regenerate an assignment from the same request
      description: Preserves requestId for traceability.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                constraints:
                  type: object
                  description: Optional updated constraints
      responses:
        "200":
          description: Regeneration started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }

  # ── Assignments ───────────────────────────────────────────────────────────
  /assignments:
    get:
      tags: [assignments]
      summary: List teacher's assignments
      responses:
        "200":
          description: Assignment list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }

    post:
      tags: [assignments]
      summary: Create assignment from a generation result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [generationResultId]
              properties:
                generationResultId: { type: string }
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
  /assignments/{id}:
    get:
      tags: [assignments]
      summary: Get assignment detail with current version content
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Assignment detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }

    patch:
      tags: [assignments]
      summary: Update assignment content (creates a new version)
      description: Editing a PUBLISHABLE assignment reverts it to DRAFT.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                instructions: { type: string }
                items:
                  type: array
                  description: Full replacement item list
      responses:
        "200":
          description: Updated (new version created)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "422":
          description: ASSIGNED assignments cannot be edited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }

    delete:
      tags: [assignments]
      summary: Delete an assignment
      description: TEACHER role required. ASSIGNED assignments cannot be deleted.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "422":
          description: Assignment is ASSIGNED and cannot be deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /assignments/{id}/publish:
    post:
      tags: [assignments]
      summary: Transition assignment to PUBLISHABLE status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Now PUBLISHABLE
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "422":
          description: Already PUBLISHABLE or ASSIGNED
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /assignments/{id}/versions:
    get:
      tags: [assignments]
      summary: List version history for an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Version list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
...
  /distribution/{id}:
    get:
      tags: [distribution]
      summary: Get distribution detail with recipient list
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Distribution detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }

  # ── Attempts ──────────────────────────────────────────────────────────────
...
  /attempts/{id}/submit:
    post:
      tags: [attempts]
      summary: Submit attempt (irreversible)
      description: STUDENT role required. Attempt becomes immutable and queues automatic AI assessment analysis.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "422":
          description: Already submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }

  # ── AI Help ───────────────────────────────────────────────────────────────
...